// frontend/src/BurialTextileRecord.interface.ts
export interface BurialTextileRecord {
    id: number;
    squarenorthsouth: string;
    headdirection: string;
    sex: string;
    northsouth: number;
    depth: number;
    eastwest: number;
    adultsubadult: string;
    facebundles: string;
    southtohead: number;
    preservation: string;
    fieldbookpage: number;
    squareeastwest: string;
    goods: string;
    text: string;
    wrapping: string;
    haircolor: string;
    westtohead: number;
    samplescollected: string;
    area: string;
    burialid: string;
    length: number;
    burialnumber: number;
    dataexpertinitials: string;
    westtofeet: number;
    ageatdeath: string;
    southtofeet: number;
    excavationrecorder: string;
    photos: string;
    hair: string;
    burialmaterials: string;
    dateofexcavation: string;
    fieldbookexcavationyear: number;
    clusternumber: number;
    shaftnumber: number;

    colorid?: number;
    description?: string;
    estimatedperiod?: string;
    locale?: string;
    photographeddate?: string;
    sampledate?: string;
    textileid?: number;
    value?: number;
};